services/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = ""
authors = [
    {name = "Nikos Zacharatos", email = "nikos.zacharatos@iccs.gr"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi>=0.128.7,<0.129.0",
    "sqlalchemy>=2.0.46,<3.0.0",
    "python-dotenv>=0.21.0",
    "pydantic>=2.12.5,<3.0.0",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "redis (>=7.2.1,<8.0.0)",
    "loguru>=0.7.2,<1.0.0",
    "python-jose[cryptography]>=3.3.0,<4.0.0",
    "bcrypt>=4.0.0,<5.0.0",
    "shared @ file:///home/nikos/home-telemetry/shared",
    "python-multipart (>=0.0.22,<0.0.23)",
]

[tool.poetry]
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
shared = {develop = true}

[tool.poetry.group.dev.dependencies]
aiosqlite = "^0.22.1"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"